requirements.txt
pytelegrambotapi
pandas
openpyxl
matplotlib
python-dotenv